finish rest api bank v0.1

+++ src/main/java/org/effectivemobile/services/BankAccountServiceImpl.java
@@ -1,5 +1,7 @@
 package org.effectivemobile.services;
 
+import org.effectivemobile.api.exceptions.ClientNotFoundException;
+import org.effectivemobile.config.ClientDetails;
 import org.effectivemobile.entity.BankAccountEntity;
 import org.effectivemobile.repositories.BankAccountRepository;
 import lombok.AccessLevel;
@@ -7,6 +9,9 @@ import lombok.AllArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.effectivemobile.repositories.ClientRepository;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,6 +24,7 @@ public class BankAccountServiceImpl implements org.effectivemobile.services.Serv
     private static final Logger logger = LogManager.getLogger(BankAccountServiceImpl.class);
 
     BankAccountRepository bankAccountRepository;
+    ClientRepository clientRepository;
 
     @Override
     public BankAccountEntity createEntity(BankAccountEntity entity) {
@@ -44,8 +50,20 @@ public class BankAccountServiceImpl implements org.effectivemobile.services.Serv
 
     @Override
     public void deleteEntityById(Long aLong) {
+        if(bankAccountRepository.findById(aLong).isEmpty() && clientRepository.findById(aLong).isEmpty()) throw new ClientNotFoundException("Not found");
         bankAccountRepository.deleteById(aLong);
+        clientRepository.deleteById(aLong);
         logger.info("Deleted bank account with ID {}", aLong);
+    }
+
+    public void deleteEntity() {
+        deleteEntityById(getId());
+    }
+
+    private Long getId() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        ClientDetails clientDetails = (ClientDetails) authentication.getPrincipal();
+        return clientDetails.getId();
     }
 
     @Override
